+++ test/lightning_web/controllers/user_session_controller_test.exs
@@ -158,7 +158,7 @@ defmodule LightningWeb.UserSessionControllerTest do
       assert redirected_to(conn) == "/"
 
       # Now do a logged in request and assert on the menu
-      conn = get(conn, "/")
+      conn = get(conn, "/projects")
       response = html_response(conn, 200)
       assert response =~ "User Profile"
     end
